override only default_filter_params from meetings controller instead of whole controller
