+++ spec/lib/overrides_spec.rb
@@ -59,6 +59,12 @@ checksums = [
       "/app/views/decidim/assemblies/_filter_by_type.html.erb" => "76988d76b84d96079e6d9e1b252a3fda",
       "/app/views/decidim/assemblies/assemblies/_parent_assemblies.html.erb" => "fd026d4ee40dd1d5ebf8ad9ec5d0dbb4"
     }
+  },
+  {
+    package: "decidim-meetings",
+    files: {
+      "/app/controllers/decidim/meetings/meetings_controller.rb" => "9b5f2673ac38699c86c89fbd0d6e326f",
+    }
   }
 ]
